# docs/openapi/course-service-openapi.yaml
openapi: 3.1.0
info:
  title: CourseService API
  description: Course content and structure service API.
  version: v1
servers:
- url: https://dev-api.studybytes.ru/course-service
  description: Development gateway through Nginx
- url: http://localhost:8082
  description: Local direct access
- url: http://course-service:8082
  description: Docker backend network
...
paths:
...
  /api/v1/internal/courses/{courseId}/ownership:
    get:
      tags:
      - Internal Course API
      summary: Check course ownership
      description: Returns whether the given user is the author of the course.
      operationId: getCourseOwnership
      parameters:
      - name: courseId
        in: path
        description: Course id
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
        example: 10
      - name: userId
        in: query
        description: User id
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
        example: 5
      responses:
        "200":
          description: Course ownership state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalCourseOwnershipResponse"
        "401":
          description: Missing or invalid internal API key
        "404":
          description: Course not found
      security:
      - internalApiKey: []
  /api/v1/internal/courses/{courseId}/availability:
    get:
      tags:
      - Internal Course API
      summary: Get course availability
      description: "Returns publication, access and enrollment state for LearningService\
        \ enrollment checks."
      operationId: getCourseAvailability
      parameters:
      - name: courseId
        in: path
        description: Course id
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: Course availability returned
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalCourseAvailabilityResponse"
        "401":
          description: Missing or invalid internal API key
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalCourseAvailabilityResponse"
        "404":
          description: Course not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalCourseAvailabilityResponse"
      security:
      - internalApiKey: []
  /api/v1/internal/course-items/{itemId}/execution-package:
    get:
      tags:
      - Internal Course API
      summary: Get course item execution package
      description: "Returns full execution data for a course item, including hidden\
        \ tests and expected outputs. Intended only for trusted backend services."
      operationId: getExecutionPackage
      parameters:
      - name: itemId
        in: path
        description: Course item id
        required: true
        schema:
          type: integer
          format: int64
        example: 2
      responses:
        "200":
          description: Execution package returned
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExecutionPackageResponse"
        "401":
          description: Missing or invalid internal API key
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExecutionPackageResponse"
        "404":
          description: Course item not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExecutionPackageResponse"
      security:
      - internalApiKey: []
  /api/v1/internal/course-items/{itemId}/content:
    get:
      tags:
      - Internal Course API
      summary: Get enrolled-student-safe course item content
      description: "Returns full course item content for trusted backend services\
        \ after enrollment checks. Does not expose hidden tests, expected outputs\
        \ or correct quiz answers."
      operationId: getCourseItemContent
      parameters:
      - name: itemId
        in: path
        description: Course item id
        required: true
        schema:
          type: integer
          format: int64
        example: 2
      responses:
        "200":
          description: Course item content returned
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalCourseItemContentResponse"
        "401":
          description: Missing or invalid internal API key
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalCourseItemContentResponse"
        "404":
          description: Course item not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalCourseItemContentResponse"
      security:
      - internalApiKey: []
...
components:
  schemas:
...
    InternalCourseOwnershipResponse:
      type: object
      properties:
        courseId:
          type: integer
          format: int64
          example: 10
        userId:
          type: integer
          format: int64
          example: 5
        owner:
          type: boolean
          example: true
      required:
      - courseId
      - owner
      - userId
    InternalCourseAvailabilityResponse:
      type: object
      properties:
        courseId:
          type: integer
          format: int64
        status:
          type: string
          enum:
          - DRAFT
          - PENDING_REVIEW
          - CHANGES_REQUESTED
          - PUBLISHED
          - ARCHIVED
        accessType:
          type: string
          enum:
          - PUBLIC
          - UNLISTED
          - PRIVATE
        enrollmentEnabled:
          type: boolean
        availableForEnrollment:
          type: boolean
    EvaluationPolicy:
      type: object
      properties:
        comparisonMode:
          type: string
          enum:
          - EXACT
        normalizeLineEndings:
          type: boolean
        trimTrailingWhitespaces:
          type: boolean
    ExecutionLimits:
      type: object
      properties:
        timeLimitMs:
          type: integer
          format: int32
        memoryLimitMb:
          type: integer
          format: int32
        outputLimitKb:
          type: integer
          format: int32
    ExecutionPackageResponse:
      type: object
      properties:
        itemId:
          type: integer
          format: int64
        moduleId:
          type: integer
          format: int64
        courseId:
          type: integer
          format: int64
        itemType:
          type: string
          enum:
          - CODING
          - SQL
          - QUIZ
          - THEORY
          - FILE
        title:
          type: string
        language:
          type: string
        starterCode:
          type: string
        limits:
          $ref: "#/components/schemas/ExecutionLimits"
        executionPolicy:
          $ref: "#/components/schemas/ExecutionPolicy"
        evaluationPolicy:
          $ref: "#/components/schemas/EvaluationPolicy"
        tests:
          type: array
          items:
            $ref: "#/components/schemas/InternalTestCaseResponse"
    ExecutionPolicy:
      type: object
      properties:
        networkDisabled:
          type: boolean
        readOnlyFs:
          type: boolean
    InternalTestCaseResponse:
      type: object
      properties:
        testKey:
          type: string
        visibility:
          type: string
          enum:
          - OPEN
          - HIDDEN
        inputData:
          type: string
        expectedOutput:
          type: string
        orderIndex:
          type: integer
          format: int32
    ContentBlockResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        blockType:
          type: string
          enum:
          - TEXT
          - VIDEO
          - IMAGE
          - CODE
          - EMBED
          - FILE
        orderIndex:
          type: integer
          format: int32
        title:
          type: string
        textContent:
          type: string
        url:
          type: string
        language:
          type: string
        metadataJson:
          type: string
    CourseItemHintResponse:
      type: object
      properties:
        orderIndex:
          type: integer
          format: int32
        text:
          type: string
    CourseItemLimitsResponse:
      type: object
      properties:
        timeLimitMs:
          type: integer
          format: int32
        memoryLimitMb:
          type: integer
          format: int32
        outputLimitKb:
          type: integer
          format: int32
    CourseItemOptionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        orderIndex:
          type: integer
          format: int32
        label:
          type: string
        text:
          type: string
    EvaluationPolicyResponse:
      type: object
      properties:
        comparisonMode:
          type: string
          enum:
          - EXACT
        normalizeLineEndings:
          type: boolean
        trimTrailingWhitespaces:
          type: boolean
    ExecutionPolicyResponse:
      type: object
      properties:
        networkDisabled:
          type: boolean
        readOnlyFs:
          type: boolean
    InternalCourseItemContentResponse:
      type: object
      properties:
        itemId:
          type: integer
          format: int64
        moduleId:
          type: integer
          format: int64
        courseId:
          type: integer
          format: int64
        title:
          type: string
        itemType:
          type: string
          enum:
          - CODING
          - SQL
          - QUIZ
          - THEORY
          - FILE
        statement:
          type: string
        starterCode:
          type: string
        language:
          type: string
        orderIndex:
          type: integer
          format: int32
        limits:
          $ref: "#/components/schemas/CourseItemLimitsResponse"
        executionPolicy:
          $ref: "#/components/schemas/ExecutionPolicyResponse"
        evaluationPolicy:
          $ref: "#/components/schemas/EvaluationPolicyResponse"
        contentBlocks:
          type: array
          items:
            $ref: "#/components/schemas/ContentBlockResponse"
        openTests:
          type: array
          items:
            $ref: "#/components/schemas/OpenTestCaseResponse"
        hints:
          type: array
          items:
            $ref: "#/components/schemas/CourseItemHintResponse"
        options:
          type: array
          items:
            $ref: "#/components/schemas/CourseItemOptionResponse"
    OpenTestCaseResponse:
      type: object
      properties:
        testKey:
          type: string
        orderIndex:
          type: integer
          format: int32
        inputData:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      description: "Bearer JWT for CourseService admin endpoints. Required roles:\
        \ TEACHER or ADMIN."
      scheme: bearer
      bearerFormat: JWT
    internalApiKey:
      type: apiKey
      description: Static service-to-service API key for internal CourseService endpoints.
        This is an MVP mechanism and should be replaced or hardened for production.
      name: X-Internal-Api-Key
      in: header